+++ backend/src/worker/mod.rs
@@ -14,9 +14,8 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-pub mod db;
+pub mod response;
 pub mod task;
 pub mod worker;
 
-pub use db::create_db;
 pub use worker::{run_worker, setup_rabbit_mq};
